Implement most of the methods of the refactored classes

+++ filedb/db.py
@@ -1,8 +1,9 @@
 import logging
-import typing
 import uuid
 from contextlib import contextmanager
+from typing import IO
 from typing import List
+from typing import Union
 
 from filedb.cache import Cache
 from filedb.index import Index
@@ -18,23 +19,39 @@ class FileDB:
                  index: Index,
                  storage: Storage,
                  cache: Cache):
-
         self.index = index
         self.storage = storage
         self.cache = cache
 
-    def find(self, query: Query) -> List[Key]:
-        return self.index.find(query)
+    def find(self, query: Query) -> List['File']:
+        return [self.file(key) for key in self.index.find(query)]
+
+    def file(self, key):
+        return File(key,
+                    index=self.index,
+                    storage=self.storage,
+                    cache=self.cache)
+
+
+class File:
+    def __init__(self,
+                 key: Key,
+                 index: Index,
+                 storage: Storage,
+                 cache: Cache):
+
+        self.key = key
+        self.index = index
+        self.storage = storage
+        self.cache = cache
 
     def read_text(self,
-                  key: Key,
                   buffering=-1,
                   encoding=None,
                   errors=None,
                   newline=None) -> str:
 
-        with self.open(key,
-                       mode="r",
+        with self.open(mode="r",
                        buffering=buffering,
                        encoding=encoding,
                        errors=errors,
@@ -42,42 +59,38 @@ class FileDB:
             return f.read()
 
     def write_text(self,
-                   key: Key,
                    data: str,
                    buffering=-1,
                    encoding=None,
                    errors=None,
                    newline=None):
 
-        with self.open(key,
-                       mode="w",
+        with self.open(mode="w",
                        buffering=buffering,
                        encoding=encoding,
                        errors=errors,
                        newline=newline) as f:
             f.write(data)
 
-    def read_bytes(self, key: Key, buffering=-1) -> bytes:
-        with self.open(key, mode="rb", buffering=buffering) as f:
+    def read_bytes(self, buffering=-1) -> bytes:
+        with self.open(mode="rb", buffering=buffering) as f:
             return f.read()
 
-    def write_bytes(self, key: Key, data: bytes, buffering=-1):
-        with self.open(key, mode="wb", buffering=buffering) as f:
+    def write_bytes(self, data: bytes, buffering=-1):
+        with self.open(mode="wb", buffering=buffering) as f:
             f.write(data)
 
     @contextmanager
     def open(self,
-             key: Key,
              mode: str = "r",
              buffering=-1,
              encoding=None,
              errors=None,
-             newline=None) -> typing.IO:
+             newline=None) -> IO:
 
         if mode[0] == "r":
 
-            with self._read_handle(key,
-                                   mode,
+            with self._read_handle(mode,
                                    buffering=buffering,
                                    encoding=encoding,
                                    errors=errors,
@@ -85,56 +98,66 @@ class FileDB:
                 yield file_object
 
         else:
-            with self._write_handle(key,
-                                    mode,
+            with self._write_handle(mode,
                                     buffering=buffering,
                                     encoding=encoding,
                                     errors=errors,
                                     newline=newline) as file_object:
                 yield file_object
 
-    def copy(self, key_1: Key, key_2: Key):
+    def copy(self, to: Union[Key, 'File']):
 
-        storage_path_1 = self.index.storage_path(key_1)
+        if isinstance(to, File) and (self.index != to.index or self.storage != to.storage):
+            raise NotImplementedError('Copying to a file in a different storage or index is not'
+                                      'yet implemented!')
+
+        to_key = to.key if isinstance(to, File) else to
+        storage_path_1 = self.index.storage_path(self.key)
         if storage_path_1 is None:
-            raise FileNotFoundError(f"File({key_1}) does not exist!")
+            raise FileNotFoundError(f"File({self.key}) does not exist!")
         storage_path_2 = uuid.uuid4()
         self.storage.copy(storage_path_1, storage_path_2)
-        self.index.upsert(key_2, storage_path_2)
+        self.index.upsert(to_key, storage_path_2)
 
-    def move(self, key_1: Key, key_2: Key):
+    def move(self, to: Union[Key, 'File']):
 
-        storage_path_1 = self.index.storage_path(key_1)
+        if isinstance(to, File) and (self.index != to.index or self.storage != to.storage):
+            raise NotImplementedError('Moving to a file in a different storage or index is not'
+                                      'yet implemented!')
+
+        to_key = to.key if isinstance(to, File) else to
+        storage_path_1 = self.index.storage_path(self.key)
         if storage_path_1 is None:
-            raise FileNotFoundError(f"File({key_1}) does not exist!")
+            raise FileNotFoundError(f"File({self.key}) does not exist!")
         storage_path_2 = uuid.uuid4()
         self.storage.copy(storage_path_1, storage_path_2)
-        self.index.upsert(key_2, storage_path_2)
-        self.index.delete(key_1)
+        self.index.upsert(to_key, storage_path_2)
+        self.index.delete(self.key)
         self.storage.delete(storage_path_1)
 
-    def delete(self, key: Key):
+    def delete(self):
 
-        storage_path = self.index.storage_path(key)
-        self.index.delete(key)
+        storage_path = self.index.storage_path(self.key)
+        self.index.delete(self.key)
         self.storage.delete(storage_path)
 
     @contextmanager
     def _read_handle(self,
-                     key: Key,
                      mode: str,
                      buffering=-1,
                      encoding=None,
                      errors=None,
                      newline=None):
 
-        storage_path = self.index.storage_path(key)
+        storage_path = self.index.storage_path(self.key)
         if storage_path is None:
-            raise FileNotFoundError(f"File({key}) does not exist!")
+            raise FileNotFoundError(f"File({self.key}) does not exist!")
 
-        cache_path = self.cache.path(storage_path, index_name=self.index.name, storage_name=self.storage.name)
+        cache_path = self.cache.path(storage_path, index_name=self.index.name,
+                                     storage_name=self.storage.name)
         with self.cache.read_lock(cache_path):
-            if not cache_path.exists() or self.cache.crc32(cache_path) != self.storage.crc32(storage_path):
+            if not cache_path.exists() or self.cache.crc32(cache_path) != self.storage.crc32(
+                    storage_path):
                 with self.cache.write_lock(cache_path):
                     self.storage.download(storage_path, cache_path)
 
@@ -147,7 +170,6 @@ class FileDB:
 
     @contextmanager
     def _write_handle(self,
-                      key: Key,
                       mode: str,
                       buffering=-1,
                       encoding=None,
@@ -155,7 +177,9 @@ class FileDB:
                       newline=None):
 
         storage_path = uuid.uuid4()
-        cache_path = self.cache.path(storage_path, index_name=self.index.name, storage_name=self.storage.name)
+        cache_path = self.cache.path(storage_path,
+                                     index_name=self.index.name,
+                                     storage_name=self.storage.name)
         with self.cache.read_lock(cache_path):
             with self.cache.write_lock(cache_path):
                 with cache_path.open(mode,
@@ -166,4 +190,4 @@ class FileDB:
                     yield f
 
             self.storage.upload(cache_path, storage_path)
-        self.index.upsert(key, storage_path)
+        self.index.upsert(self.key, storage_path)
